Make report add a server action
